requirements.txt
aiogram>=3.4.0
qrcode>=7.4.2
Pillow>=10.2.0
python-dotenv>=1.0.1
APScheduler>=3.10.4
SQLAlchemy>=2.0.27
aiosqlite>=0.20.0
httpx>=0.27.0
